# services/agents-api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Agents API",
    "description" : "Conversation and messaging service for jorisjonkers.dev",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://agents.jorisjonkers.dev",
    "description" : "Production"
  }, {
    "url" : "http://localhost:8082",
    "description" : "Local development"
  } ],
  "paths" : {
...
    "/api/v1/conversations" : {
      "get" : {
        "tags" : [ "conversation-controller" ],
        "operationId" : "listByUser",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ConversationResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "conversation-controller" ],
        "operationId" : "create_3",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateConversationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConversationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/conversations/{conversationId}/messages" : {
      "get" : {
        "tags" : [ "message-controller" ],
        "operationId" : "list_3",
        "parameters" : [ {
          "name" : "conversationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/MessageResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "message-controller" ],
        "operationId" : "send_1",
        "parameters" : [ {
          "name" : "conversationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat-sessions" : {
      "get" : {
        "tags" : [ "chat-session-controller" ],
        "operationId" : "list_4",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ChatSessionResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "chat-session-controller" ],
        "operationId" : "create_4",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StartChatSessionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChatSessionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat-sessions/{id}/messages" : {
      "post" : {
        "tags" : [ "chat-session-controller" ],
        "operationId" : "appendMessage",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AppendChatMessageRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/credentials/sessions/{id}" : {
      "get" : {
        "tags" : [ "credential-controller" ],
        "summary" : "Get the current status of a re-authentication session",
        "operationId" : "status",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/conversations/{id}" : {
      "get" : {
        "tags" : [ "conversation-controller" ],
        "operationId" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConversationResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "conversation-controller" ],
        "operationId" : "archive",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
    "/api/v1/chat-sessions/{id}" : {
      "get" : {
        "tags" : [ "chat-session-controller" ],
        "operationId" : "get_3",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "chat-session-controller" ],
        "operationId" : "archive_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/v1/workspaces/{workspaceId}/sessions/{sessionId}" : {
      "delete" : {
        "tags" : [ "agent-session-controller" ],
        "operationId" : "stop",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "sessionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CreateConversationRequest" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "maxLength" : 200,
            "minLength" : 0
          }
        },
        "required" : [ "title" ]
      },
      "ConversationResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "userId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "createdAt", "id", "status", "title", "updatedAt", "userId" ]
      },
      "SendMessageRequest" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "string",
            "maxLength" : 10000,
            "minLength" : 0
          }
        },
        "required" : [ "content" ]
      },
      "MessageResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "conversationId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "role" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "content", "conversationId", "createdAt", "id", "role" ]
      },
      "StartChatSessionRequest" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : [ "string", "null" ],
            "maxLength" : 120,
            "minLength" : 0
          },
          "kind" : {
            "type" : [ "string", "null" ],
            "enum" : [ "PLAIN", "KNOWLEDGE" ]
          }
        }
      },
      "ChatSessionResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "userId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "type" : "string"
          },
          "kind" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "createdAt", "id", "kind", "status", "updatedAt", "userId" ]
      },
      "AppendChatMessageRequest" : {
        "type" : "object",
        "properties" : {
          "body" : {
            "type" : "string",
            "minLength" : 1
          },
          "role" : {
            "type" : "string",
            "enum" : [ "USER", "ASSISTANT", "SYSTEM" ]
          }
        },
        "required" : [ "body", "role" ]
      },
      "ChatMessageResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "sessionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "role" : {
            "type" : "string"
          },
          "body" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "body", "createdAt", "id", "role", "sessionId" ]
      },
...
    }
  }
}